test(range): migrate a11y
